+++ packages/service-library/src/servicelib/fastapi/http_error.py
@@ -1,30 +1,14 @@
 from collections.abc import Awaitable, Callable
 from typing import TypeVar
 
-from fastapi import HTTPException
+from fastapi import FastAPI, HTTPException, status
 from fastapi.encoders import jsonable_encoder
+from fastapi.exceptions import RequestValidationError
 from fastapi.openapi.constants import REF_PREFIX
 from fastapi.openapi.utils import validation_error_response_definition
 from fastapi.requests import Request
 from fastapi.responses import JSONResponse
-
-
-def make_default_http_error_handler(
-    *, envelope_error: bool
-) -> Callable[[Request, Exception], Awaitable[JSONResponse]]:
-    async def _http_error_handler(_: Request, exc: Exception) -> JSONResponse:
-        assert isinstance(exc, HTTPException)
-
-        error_content = {"errors": [exc.detail]}
-        if envelope_error:
-            error_content = {"error": error_content}
-        return JSONResponse(
-            content=jsonable_encoder(error_content),
-            status_code=exc.status_code,
-        )
-
-    return _http_error_handler
-
+from pydantic import ValidationError
 
 validation_error_response_definition["properties"] = {
     "errors": {
@@ -66,3 +50,51 @@ def make_http_error_handler_for_exception(
         )
 
     return _http_error_handler
+
+
+def _request_validation_error_extractor(
+    validation_error: RequestValidationError,
+) -> list[str]:
+    return [f"{e}" for e in validation_error.errors()]
+
+
+def _make_default_http_error_handler(
+    *, envelope_error: bool
+) -> Callable[[Request, Exception], Awaitable[JSONResponse]]:
+    async def _http_error_handler(_: Request, exc: Exception) -> JSONResponse:
+        assert isinstance(exc, HTTPException)
+
+        error_content = {"errors": [exc.detail]}
+        if envelope_error:
+            error_content = {"error": error_content}
+        return JSONResponse(
+            content=jsonable_encoder(error_content),
+            status_code=exc.status_code,
+        )
+
+    return _http_error_handler
+
+
+def set_app_default_http_error_handlers(app: FastAPI) -> None:
+    app.add_exception_handler(
+        HTTPException, _make_default_http_error_handler(envelope_error=True)
+    )
+
+    app.add_exception_handler(
+        RequestValidationError,
+        make_http_error_handler_for_exception(
+            status.HTTP_422_UNPROCESSABLE_ENTITY,
+            RequestValidationError,
+            envelope_error=True,
+            error_extractor=_request_validation_error_extractor,
+        ),
+    )
+
+    app.add_exception_handler(
+        ValidationError,
+        make_http_error_handler_for_exception(
+            status.HTTP_422_UNPROCESSABLE_ENTITY,
+            ValidationError,
+            envelope_error=True,
+        ),
+    )
